+++ lib/instream/connection/query_planner.ex
@@ -3,17 +3,33 @@ defmodule Instream.Connection.QueryPlanner do
   Query planning coordinating the execution of all queries.
   """
 
+  alias Instream.Data.Read
+  alias Instream.Encoder.InfluxQL
   alias Instream.Query
+  alias Instream.Query.Builder
 
   @doc """
   Executes a query.
   """
-  @spec execute(Query.t, Keyword.t, module) :: any
+  @spec execute(Builder.t | Query.t | String.t, Keyword.t, module) :: any
+  def execute(%Builder{} = query, opts, conn) do
+    query
+    |> InfluxQL.encode()
+    |> Read.query(opts)
+    |> execute(opts, conn)
+  end
+
   def execute(%Query{} = query, opts, conn) do
     case opts[:async] do
       true -> execute_async(query, opts, conn)
       _    -> execute_sync(query, opts, conn)
     end
+  end
+
+  def execute(query, opts, conn) when is_binary(query) do
+    query
+    |> Read.query(opts)
+    |> execute(opts, conn)
   end